# 6_express/tanfolyam-express-8/src/controllers/post/post.service.js
const Post = require('../../models/post.model');
const Person = require('../../models/person.model')

exports.create = postData => {
  const post = new Post(postData);
  return post.save()
    .then(() => Person.findById(postData.author))
    .then(author => {
      author.posts.push(post._id);
      return author.save
    })
    .then(() => post);
}

// kikeresünk egy post-ot és populate metódussal hozzárendeljük az authort
exports.findOne = id => Post.findById(id).populate('author')