# 6_express/tanfolyam-express-9/docs/swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Sample Blog API
  description: A sample blog API to illustrate REST API concepts
servers:
  - url: http://localhost:3000
paths:
  /person:
    get:
      description: Returns a list of persons
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of persons returned successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Person"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      description: Creates a new person
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Person"
      responses:
        "201":
          description: Person has been created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /person/{person_id}:
    get:
      description: Returns the person with the given ID
      security:
        - bearerAuth: []
      parameters:
        - name: person_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Person data returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
        "404":
          description: Person with the give ID was not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      description: Update a person
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Person"
      responses:
        "201":
          description: Person has been updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Person:
      type: object
      required:
        - first_name
        - last_name
        - email
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
    Post:
      type: object
      required:
        - title
        - body
      properties:
        title:
          type: string
        body:
          type: string
        author:
          type: object
          $ref: "#/components/schemas/Person"
    Error:
      type: object
      required:
        - message
        - has_error
      properties:
        message:
          type: string
        has_error:
          type: boolean